refactor: encapsulate logic
Encapsulation based on roles
Simplify main() in application.kt
